python scripts cleanup
