+++ .github/workflows/scripts/release-canister.py
@@ -1,32 +1,16 @@
-from ic.client import Client
-from ic.identity import Identity
-from ic.agent import Agent
 import os
 import sys
-import base64
-from ic.candid import encode, Types
 
 if len(sys.argv) != 2:
-  print("Usage: python3 request_canister.py <ref>")
+  print("Usage: python3 release-canister.py <ref>")
   exit(1)
 
 for v in ["DFX_IDENTITY_PREVIEW","POOL_CANISTER_ID"]:
   if not v in os.environ:
-    print(f"request_canister.py: {v} env variable missing")
+    print(f"release-canister.py: {v} env variable missing")
     exit(1)
 
-private_key = base64.b64decode(os.environ["DFX_IDENTITY_PREVIEW"]).decode("utf-8")
-pool_id = os.environ["POOL_CANISTER_ID"]
 
-identity = Identity.from_pem(private_key)
-client = Client()
-agent = Agent(identity, client)
-
-
-def release_canister():
-  res = agent.update_raw(
-      pool_id, "release_canister", encode([{'type': Types.Text, 'value': sys.argv[1]}]))
-  return res
-
+from pool import release_canister
 
 release_canister()
